+++ config.php
@@ -23,6 +23,7 @@ if (file_exists("./configdatabase.txt") == false) { // If the database file does
     $config["backup_overwriting"] = false;
     $config["auto_backup"] = "";
     $config["auto_backup_interval"] = 60;
+    $config["default_max_items"] = 2048;
     file_put_contents("./configdatabase.txt", serialize($config)); // Write the configuration database to disk.
 
 } else { // Otherwise, the file exists, so load the configuration database from disk.