+++ configuration.php
@@ -18,8 +18,9 @@ $credit_level = $_POST["creditlevel"]; // This is the level of credit given to V
 $display_advanced_tools = $_POST["displayadvancedtools"]; // This determines whether or not the advanced tools section will be displayed on the tools page.
 $displayed_search_results_count = $_POST["displayedsearchresultscount"]; // This is the number of results that will be displayed in the Search tool.
 $backup_overwriting = $_POST["backupoverwriting"]; // This determines whether the database backup tool will allow the user to overwrite existing files.
-$auto_backup = $_POST["autobackup"];
-$auto_backup_interval = $_POST["autobackupinterval"];
+$auto_backup = $_POST["autobackup"]; // This determines where automatic backups will be created.
+$auto_backup_interval = $_POST["autobackupinterval"]; // This determines how often automatic backups can be created.
+$default_max_items = $_POST["defaultmaxitems"]; // This determines the maximum number of items that a user can have in their database by default.
 
 
 
@@ -110,9 +111,12 @@ if ($auto_backup != filter_var($auto_backup, FILTER_SANITIZE_STRING)) {
 
 $auto_backup_interval = intval($auto_backup_interval);
 
+$default_max_items = intval($default_max_items);
 
 
 
+
+// Convert the raw whitelist input string into a list array.
 $whitelist = explode(",", $whitelist);
 foreach ($whitelist as $key => $user) { // Iterate through all users in the list of whitelisted users.
     $whitelist[$key] = trim($user); // Trim any leading or trailing blank spaces for each user.
@@ -145,6 +149,7 @@ if ($theme != null) { // Check to see if information was input through the form.
     $config["backup_overwriting"] = $backup_overwriting;
     $config["auto_backup"] = $auto_backup;
     $config["auto_backup_interval"] = $auto_backup_interval;
+    $config["default_max_items"] = $default_max_items;
 
 
     file_put_contents("./configdatabase.txt", serialize($config)); // Write database changes to disk.
@@ -222,6 +227,8 @@ $formatted_whitelist = substr($formatted_whitelist, 1); // Remove the first char
             <label for="autobackup">Auto-Backup Path: </label><input id="autobackup" name="autobackup" type="text" value="<?php echo $config["auto_backup"]; ?>" placeholder="/home/user/BackupFile">
             <br><br>
             <label for="autobackupinterval">Auto-Backup Interval: </label><input id="autobackupinterval" name="autobackupinterval" type="number" step="0" min="1" value="<?php echo $config["auto_backup_interval"]; ?>" placeholder="60 seconds">
+            <br><br>
+            <label for="defaultmaxitems">Default Max Items: </label><input id="defaultmaxitems" name="defaultmaxitems" type="number" step="1" min="1" value="<?php echo $config["default_max_items"]; ?>" placeholder="1000 items">
             <br><br>
             <input type="submit" value="Submit">
         </form>